# web/src/lib/api/openapi.json
{
  "openapi": "3.1.0",
  "info": {
    "title": "Understudy",
    "description": "Rehearse-then-dial voice recruiting on the Hunar Voice Agents API.",
    "version": "0.1.0"
  },
  "paths": {
...
    "/candidates/{candidate_id}": {
      "patch": {
        "tags": [
          "candidates"
        ],
        "summary": "Update a candidate's phone, language, or DNC flag",
        "description": "A plain field edit \u2014 unlike POST .../consent, this never sets consent_recorded_at. The outbound-call guard still checks consent independently at call time regardless of what phone_e164 is set to here.",
        "operationId": "patch_candidate_candidates__candidate_id__patch",
        "parameters": [
          {
            "name": "candidate_id",
            "in": "path",
            "required": true,
            "schema": {
              "type": "string",
              "format": "uuid",
              "title": "Candidate Id"
            }
          }
        ],
        "requestBody": {
          "required": true,
          "content": {
            "application/json": {
              "schema": {
                "$ref": "#/components/schemas/CandidatePatch"
              }
            }
          }
        },
        "responses": {
          "200": {
            "description": "Successful Response",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/CandidateRead"
                }
              }
            }
          },
          "422": {
            "description": "Validation Error",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/HTTPValidationError"
                }
              }
            }
          }
        }
      }
    },
    "/candidates/{candidate_id}/consent": {
      "post": {
        "tags": [
          "candidates"
        ],
        "summary": "Record consent",
        "description": "Sets phone_e164 and consent_recorded_at together \u2014 the only route that does, per CLAUDE.md's rule that a call needs an explicitly consented number.",
        "operationId": "record_candidate_consent_candidates__candidate_id__consent_post",
        "parameters": [
          {
            "name": "candidate_id",
            "in": "path",
            "required": true,
            "schema": {
              "type": "string",
              "format": "uuid",
              "title": "Candidate Id"
            }
          }
        ],
        "requestBody": {
          "required": true,
          "content": {
            "application/json": {
              "schema": {
                "$ref": "#/components/schemas/ConsentCreate"
              }
            }
          }
        },
        "responses": {
          "200": {
            "description": "Successful Response",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/CandidateRead"
                }
              }
            }
          },
          "422": {
            "description": "Validation Error",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/HTTPValidationError"
                }
              }
            }
          }
        }
      }
    },
...
  },
  "components": {
    "schemas": {
...
      "CandidatePatch": {
        "properties": {
          "phone_e164": {
            "anyOf": [
              {
                "type": "string"
              },
              {
                "type": "null"
              }
            ],
            "title": "Phone E164"
          },
          "preferred_language": {
            "anyOf": [
              {
                "$ref": "#/components/schemas/Language"
              },
              {
                "type": "null"
              }
            ]
          },
          "dnc": {
            "anyOf": [
              {
                "type": "boolean"
              },
              {
                "type": "null"
              }
            ],
            "title": "Dnc"
          }
        },
        "additionalProperties": false,
        "type": "object",
        "title": "CandidatePatch",
        "description": "PATCH /candidates/{id} \u2014 phone, language, dnc, per the API surface. Every field is\noptional; only what is provided is changed."
      },
      "CandidateRead": {
        "properties": {
          "id": {
            "type": "string",
            "format": "uuid",
            "title": "Id"
          },
          "job_id": {
            "type": "string",
            "format": "uuid",
            "title": "Job Id"
          },
          "source_provider": {
            "type": "string",
            "title": "Source Provider"
          },
          "source_ref": {
            "type": "string",
            "title": "Source Ref"
          },
          "full_name": {
            "type": "string",
            "title": "Full Name"
          },
          "headline": {
            "anyOf": [
              {
                "type": "string"
              },
              {
                "type": "null"
              }
            ],
            "title": "Headline"
          },
          "current_title": {
            "anyOf": [
              {
                "type": "string"
              },
              {
                "type": "null"
              }
            ],
            "title": "Current Title"
          },
          "current_company": {
            "anyOf": [
              {
                "type": "string"
              },
              {
                "type": "null"
              }
            ],
            "title": "Current Company"
          },
          "location": {
            "anyOf": [
              {
                "type": "string"
              },
              {
                "type": "null"
              }
            ],
            "title": "Location"
          },
          "skills": {
            "items": {},
            "type": "array",
            "title": "Skills"
          },
          "years_experience": {
            "anyOf": [
              {
                "type": "number"
              },
              {
                "type": "null"
              }
            ],
            "title": "Years Experience"
          },
          "linkedin_url": {
            "anyOf": [
              {
                "type": "string"
              },
              {
                "type": "null"
              }
            ],
            "title": "Linkedin Url"
          },
          "phone_e164": {
            "anyOf": [
              {
                "type": "string"
              },
              {
                "type": "null"
              }
            ],
            "title": "Phone E164"
          },
          "preferred_language": {
            "anyOf": [
              {
                "$ref": "#/components/schemas/Language"
              },
              {
                "type": "null"
              }
            ]
          },
          "match_score": {
            "anyOf": [
              {
                "type": "number"
              },
              {
                "type": "null"
              }
            ],
            "title": "Match Score"
          },
          "match_breakdown": {
            "anyOf": [
              {
                "additionalProperties": true,
                "type": "object"
              },
              {
                "type": "null"
              }
            ],
            "title": "Match Breakdown"
          },
          "consent_recorded_at": {
            "anyOf": [
              {
                "type": "string",
                "format": "date-time"
              },
              {
                "type": "null"
              }
            ],
            "title": "Consent Recorded At"
          },
          "consent_channel": {
            "anyOf": [
              {
                "type": "string"
              },
              {
                "type": "null"
              }
            ],
            "title": "Consent Channel"
          },
          "dnc": {
            "type": "boolean",
            "title": "Dnc"
          }
        },
        "additionalProperties": false,
        "type": "object",
        "required": [
          "id",
          "job_id",
          "source_provider",
          "source_ref",
          "full_name",
          "headline",
          "current_title",
          "current_company",
          "location",
          "skills",
          "years_experience",
          "linkedin_url",
          "phone_e164",
          "preferred_language",
          "match_score",
          "match_breakdown",
          "consent_recorded_at",
          "consent_channel",
          "dnc"
        ],
        "title": "CandidateRead"
      },
...
      "ConsentCreate": {
        "properties": {
          "phone_e164": {
            "type": "string",
            "minLength": 1,
            "title": "Phone E164"
          },
          "channel": {
            "type": "string",
            "title": "Channel",
            "default": "MANUAL"
          }
        },
        "additionalProperties": false,
        "type": "object",
        "required": [
          "phone_e164"
        ],
        "title": "ConsentCreate"
      },
      "HTTPValidationError": {
        "properties": {
          "detail": {
            "items": {
              "$ref": "#/components/schemas/ValidationError"
            },
            "type": "array",
            "title": "Detail"
          }
        },
        "type": "object",
        "title": "HTTPValidationError"
      },
...
      "Language": {
        "type": "string",
        "enum": [
          "ENGLISH",
          "HINDI",
          "TAMIL",
          "TELUGU",
          "KANNADA",
          "MARATHI",
          "MALAYALAM",
          "GUJARATI",
          "BENGALI",
          "TURKISH",
          "ARABIC",
          "SPANISH"
        ],
        "title": "Language",
        "description": "Belongs to the agent, not the call \u2014 see CLAUDE.md."
      },
...
      "ValidationError": {
        "properties": {
          "loc": {
            "items": {
              "anyOf": [
                {
                  "type": "string"
                },
                {
                  "type": "integer"
                }
              ]
            },
            "type": "array",
            "title": "Location"
          },
          "msg": {
            "type": "string",
            "title": "Message"
          },
          "type": {
            "type": "string",
            "title": "Error Type"
          },
          "input": {
            "title": "Input"
          },
          "ctx": {
            "type": "object",
            "title": "Context"
          }
        },
        "type": "object",
        "required": [
          "loc",
          "msg",
          "type"
        ],
        "title": "ValidationError"
      },
...
    }
  },
...
}